# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CounterClaimIn": {
        "properties": {
          "counter_claim": {
            "maxLength": 2000,
            "minLength": 12,
            "title": "Counter Claim",
            "type": "string"
          }
        },
        "required": [
          "counter_claim"
        ],
        "title": "CounterClaimIn",
        "type": "object"
      },
...
      "DisputeIn": {
        "properties": {
          "claim": {
            "maxLength": 2000,
            "minLength": 12,
            "title": "Claim",
            "type": "string"
          }
        },
        "required": [
          "claim"
        ],
        "title": "DisputeIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HumanReviewIn": {
        "properties": {
          "action": {
            "enum": [
              "APPROVE",
              "REJECT"
            ],
            "title": "Action",
            "type": "string"
          },
          "reason": {
            "maxLength": 1000,
            "minLength": 12,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "action",
          "reason"
        ],
        "title": "HumanReviewIn",
        "type": "object"
      },
...
      "ResolveDisputeIn": {
        "properties": {
          "reason": {
            "maxLength": 1000,
            "minLength": 12,
            "title": "Reason",
            "type": "string"
          },
          "refund_paise": {
            "minimum": 0.0,
            "title": "Refund Paise",
            "type": "integer"
          },
          "release_paise": {
            "minimum": 0.0,
            "title": "Release Paise",
            "type": "integer"
          }
        },
        "required": [
          "release_paise",
          "refund_paise",
          "reason"
        ],
        "title": "ResolveDisputeIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Programmable escrow for agentic commerce. The LLM never moves money: it writes a signed attestation, and a deterministic settlement engine reads it.",
    "title": "Aegis",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/disputes/{dispute_id}": {
      "get": {
        "operationId": "get_dispute_api_v1_disputes__dispute_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "dispute_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dispute Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Dispute Api V1 Disputes  Dispute Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Dispute",
        "tags": [
          "disputes"
        ]
      }
    },
    "/api/v1/disputes/{dispute_id}/arbiter": {
      "post": {
        "description": "Runs the advisory arbiter.  It cannot move money (I8).",
        "operationId": "run_arbiter_api_v1_disputes__dispute_id__arbiter_post",
        "parameters": [
          {
            "in": "path",
            "name": "dispute_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dispute Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Run Arbiter Api V1 Disputes  Dispute Id  Arbiter Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Arbiter",
        "tags": [
          "disputes"
        ]
      }
    },
    "/api/v1/disputes/{dispute_id}/counter-claim": {
      "post": {
        "operationId": "counter_claim_api_v1_disputes__dispute_id__counter_claim_post",
        "parameters": [
          {
            "in": "path",
            "name": "dispute_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dispute Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CounterClaimIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Counter Claim Api V1 Disputes  Dispute Id  Counter Claim Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Counter Claim",
        "tags": [
          "disputes"
        ]
      }
    },
    "/api/v1/disputes/{dispute_id}/resolve": {
      "post": {
        "operationId": "resolve_api_v1_disputes__dispute_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "dispute_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dispute Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveDisputeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Resolve Api V1 Disputes  Dispute Id  Resolve Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve",
        "tags": [
          "disputes"
        ]
      }
    },
...
    "/api/v1/milestones/{milestone_id}/disputes": {
      "post": {
        "operationId": "raise_dispute_api_v1_milestones__milestone_id__disputes_post",
        "parameters": [
          {
            "in": "path",
            "name": "milestone_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Milestone Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Raise Dispute Api V1 Milestones  Milestone Id  Disputes Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Raise Dispute",
        "tags": [
          "disputes"
        ]
      }
    },
    "/api/v1/milestones/{milestone_id}/human-review": {
      "post": {
        "description": "APPROVE or REJECT an escalated milestone.  A written reason is mandatory.",
        "operationId": "human_review_api_v1_milestones__milestone_id__human_review_post",
        "parameters": [
          {
            "in": "path",
            "name": "milestone_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Milestone Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HumanReviewIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Human Review Api V1 Milestones  Milestone Id  Human Review Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Human Review",
        "tags": [
          "disputes"
        ]
      }
    },
...
  }
}